# perf: force index usage in `api/v2/addresses/:hash/transactions`